requirements.txt
pymupdf
google-genai
python-dotenv
rich
tenacity
pytesseract
Pillow
pandas
